+++ webservice/src/main/java/com/hortonworks/iotas/webservice/MetricsResource.java
@@ -1,0 +1,67 @@
+package com.hortonworks.iotas.webservice;
+
+import com.codahale.metrics.annotation.Timed;
+import com.hortonworks.iotas.catalog.Topology;
+import com.hortonworks.iotas.common.IotasEvent;
+import com.hortonworks.iotas.notification.common.Notification;
+import com.hortonworks.iotas.notification.service.NotificationService;
+import com.hortonworks.iotas.service.CatalogService;
+import com.hortonworks.iotas.topology.TopologyMetrics;
+import com.hortonworks.iotas.webservice.util.WSUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.PUT;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.MultivaluedMap;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.UriInfo;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+
+import static com.hortonworks.iotas.catalog.CatalogResponse.ResponseMessage.ENTITY_NOT_FOUND;
+import static com.hortonworks.iotas.catalog.CatalogResponse.ResponseMessage.ENTITY_NOT_FOUND_FOR_FILTER;
+import static com.hortonworks.iotas.catalog.CatalogResponse.ResponseMessage.EXCEPTION;
+import static com.hortonworks.iotas.catalog.CatalogResponse.ResponseMessage.SUCCESS;
+import static javax.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
+import static javax.ws.rs.core.Response.Status.OK;
+
+/**
+ * REST endpoint for metrics
+ */
+@Path("/api/v1/metrics")
+@Produces(MediaType.APPLICATION_JSON)
+public class MetricsResource {
+    private static final Logger LOG = LoggerFactory.getLogger(MetricsResource.class);
+
+    private CatalogService catalogService;
+
+    public MetricsResource(CatalogService service) {
+        this.catalogService = service;
+    }
+
+    @GET
+    @Path("/topologies/{id}")
+    @Timed
+    public Response getTopologyMetricsById(@PathParam("id") Long id) {
+        try {
+            Topology topology = catalogService.getTopology(id);
+            if (topology != null) {
+                Map<String, TopologyMetrics.ComponentMetric> topologyMetrics = catalogService.getTopologyMetrics(topology);
+                return WSUtils.respond(OK, SUCCESS, topologyMetrics);
+            }
+        } catch (Exception ex) {
+            LOG.error("Got exception", ex);
+            return WSUtils.respond(INTERNAL_SERVER_ERROR, EXCEPTION, ex.getMessage());
+        }
+        return WSUtils.respond(NOT_FOUND, ENTITY_NOT_FOUND, id.toString());
+    }
+}